+++ CHANGELOG.md
@@ -8,7 +8,23 @@ and this project adheres to [Semantic Versioning][].
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## [0.1.0] - tbd
+## [0.2.1] - 2024-xx-xx
+
+### Fixed
+
+-   Fixed `fill_alpha` ignoring `alpha` channel from custom cmap
+
+## [0.2.0] - 2024-03-24
+
+### Added
+
+-   Support for plotting multiple tables @melonora
+
+### Fixed
+
+-   Several bugfixes, especially for colors and palettes @melonora
+
+## [0.1.0] - 2024-01-17
 
 ### Added